Changes to start_date
Changes start_date to be 7 days previous to the last successful run of the flow. This is to account for some delay in the data usually caused by the meters themselves being set to the incorrect date and/or time. Running this flow will take 7 minutes, which I hope is OK.